Refactor base tests so test classes importable elsewhere
